Remove bloom filter implementation
This was an experimental filter that was supposed to improve
memory usage with large blocklists. There seems to be a lot of
overhead when it comes to LUA, storing objects in memory and as
a result, the bloom filter ends up taking roughly the same amount
of RAM as just storing the strings in a table.

There are also bugs with the current version of the Bloom Filter
used, but as the implementation is not memory-efficient, it's
left as-is.