Add experimental (ffi) filter
Add a simple wrapper around a binary fuse filter, which allows
to call it from Lua via ffi.

This should (in theory) provide a faster and more efficient
implementation while keeping away from Lua objects and structures.

The current implementation keeps the resulting filter as a
variable in Lua, this simplifies the setup, but it may not be as
efficient. The upside is that we avoid dealing with a state-machine.

Should scale and false positives should be minimal, though more
Testing is required.

Due to the probabilistic nature of the filter, there may be
false positives.

It's considered experimental and may not work as expected, if
at all.